last commit

+++ database/migrations/2024_04_29_111200_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string("name");
             $table->string("description");
             $table->decimal('price');
-            $table->decimal('stock');
+            $table->integer('stock');
             $table->string('image');
             $table->timestamps();
         });
